// src/models/common/MechComponentType.ts
export enum MechComponentType {
  Head = 'Head',
  CenterTorso = 'Torso (Center)',
  LeftTorso = 'Torso (Left)',
  RightTorso = 'Torso (Right)',
  LeftArm = 'Left Arm',
  RightArm = 'Right Arm',
  LeftLeg = 'Left Leg',
  RightLeg = 'Right Leg',
  FrontLeftLeg = 'Front Left Leg',
  FrontRightLeg = 'Front Right Leg',
  RearLeftLeg = 'Rear Left Leg',
  RearRightLeg = 'Rear Right Leg',
}